kucoin/requirements.txt.txt
kucoin-python
python-dotenv
pandas
requests